% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naspaclust.R
\name{ayam}
\alias{ayam}
\title{Add together two numbers}
\usage{
ayam(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
ayam(1, 1)
babi(10, 1)

}
