% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census2010dist}
\alias{census2010dist}
\title{Indonesia Provincial Matrix Distance}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 33 rows and 33 columns.
}
\usage{
census2010dist
}
\description{
A dataset containing the calculated distance between provinces in Indonesia.
}
\details{
A matrix with 33x33 rows.\cr
The dataset was firstly used in \insertCite{Wijayanto2015a;textual}{naspaclust}, \insertCite{fgwcpso;textual}{naspaclust}, \insertCite{fgwcabc1;textual}{naspaclust}, and \insertCite{fgwcabc2;textual}{naspaclust}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
