% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census2010pop}
\alias{census2010pop}
\title{Indonesia 2010 Population}
\format{
An object of class \code{integer} of length 33.
}
\usage{
census2010pop
}
\description{
A dataset containing the 2010 population in Indonesia based on census result.
}
\details{
A vector with 33 elements.\cr
The dataset was firstly used in \insertCite{Wijayanto2015a;textual}{naspaclust}, \insertCite{fgwcpso;textual}{naspaclust}, \insertCite{fgwcabc1;textual}{naspaclust}, and \insertCite{fgwcabc2;textual}{naspaclust}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
