% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei.R
\name{eiDist}
\alias{eiDist}
\alias{logchaotic}
\alias{kentchaotic}
\alias{update_alpha}
\title{Noise distributions}
\usage{
eiDist(
  distr = "normal",
  n,
  randomN = 40,
  r = 4,
  m = 0.7,
  ind = 1,
  skew = 0,
  sca = 1
)

logchaotic(n, r = 4, seed = 1)

kentchaotic(n, m = 0.7, seed)

update_alpha(alpha, iter, maxiter, type)
}
\arguments{
\item{distr}{the distribution between \code{"normal"}, \code{"uniform"}, \code{"levy"}, \code{logchaotic}, \code{kentchaotic}}

\item{n}{number of observations.}

\item{randomN}{random seed for initialisation.}

\item{r}{weight in logistic chaotic between [0,4]. Can be used when \code{ei.distr='logchaotic'}. Can be defined as \code{chaos} in \code{opt_param}. Default is 4.}

\item{m}{mapping parameter in kent chaotic between [0,1]. Can be used when \code{ei.distr='kentchaotic'}. Can be defined as \code{map} in \code{opt_param}. Default is 0.7.}

\item{ind}{Levy distribution index for random walk. Can be used when \code{ei.distr='levy'}. Can be defined as \code{ind} in \code{opt_param}. Default is 1.}

\item{skew}{Levy distribution skewness for random walk. Can be used when \code{ei.distr='levy'}. Can be defined as \code{skew} in \code{opt_param}. Default is 0.}

\item{sca}{Levy distribution scale for random walk. Can be used when \code{ei.distr='levy'}. Can be defined as \code{sca} in \code{opt_param}. Default is 1.}

\item{seed}{the random number}

\item{alpha}{the current position of alpha}

\item{iter}{the current iteration}

\item{maxiter}{the maximum iteration}

\item{type}{the update type}
}
\description{
Generating various distributions for noise in optimisation algorithm
}
