% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_fgwc}
\alias{index_fgwc}
\alias{PC1}
\alias{CE1}
\alias{SC1}
\alias{SI1}
\alias{XB1}
\alias{IFV1}
\alias{Kwon1}
\title{Fuzzy Clustering indices}
\usage{
index_fgwc(data, cluster, uij, vi, m, a = exp(1))

PC1(uij)

CE1(uij, a = exp(1))

SC1(data, cluster, uij, vi, m)

SI1(data, uij, vi)

XB1(data, uij, vi, m)

IFV1(data, uij, vi, m)

Kwon1(data, uij, vi, m)
}
\arguments{
\item{data}{an object of data with d>1. Can be \code{matrix} or \code{data.frame}. If your data is univariate, bind it with \code{1} to get a 2 columns.}

\item{cluster}{a vector. The Clustering results.}

\item{uij}{membership matrix.}

\item{vi}{centroid matrix.}

\item{m}{fuzzifier}

\item{a}{the logarithm base for CE index}
}
\description{
Fuzzy Clustering Index calculation using various literatures
}
