% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census2010}
\alias{census2010}
\title{Indonesia 2010 Provincial Census data}
\format{
An object of class \code{data.frame} with 33 rows and 110 columns.
}
\usage{
census2010
}
\description{
A dataset containing the demographic characteristics of Indonesia on provincial level
from BPS-Statistics Indonesia
}
\details{
A data frame with 33 rows and 110 demographic characteristics.\cr
The dataset was used in \insertCite{fgwcpso}{naspaclust}, \insertCite{fgwcabc1}{naspaclust} and \insertCite{fgwcabc2}{naspaclust}.\cr
The details of each variable can be seen in \insertCite{Wijayanto2015a}{naspaclust},
}
\references{
\insertAllCited{}
}
\keyword{datasets}
