% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census2010pop}
\alias{census2010pop}
\title{Indonesia 2010 Population}
\format{
A vector with 33 elements.\cr
The dataset was firstly used in Wijayanto (2014), Wijayanto and Purwarianti (2014) and Wijayanto et al. (2015)
}
\source{
\itemize{
   \item BPS, Statistics Indonesia - 2010 Population Census, Available: sp2010.bps.go.id.
   \item Wijayanto, Arie Wahyu. (2015). 
         Improvement of Fuzzy Geo-Demographic Clustering using Metaheuristic Optimization on Indonesia Population Census. 
   \item Wijayanto, A.W., Purwarianti, A., Son, L.H., 
   Fuzzy geographically weighted clustering using artificial bee colony: An efficient geo-demographic 
   analysis algorithm and applications to the analysis of crime behavior in population. 
   Applied Intelligence 44(2), 377–398 (2015).
   \item Wijayanto, A.W., Purwarianti, A., 
   Improvement design of fuzzy geo-demographic clustering using artificial bee colony optimization. 
   In: 2014 International Conference on Cyber and IT Service Management (CITSM), 1, pp. 69–74. IEEE (2014).
}
}
\usage{
census2010pop
}
\description{
A dataset containing the 2010 population in Indonesia based on census result.
}
\keyword{datasets}
